+++ extract_msg/__init__.py
@@ -27,8 +27,8 @@ https://github.com/TeamMsgExtractor/msg-extractor
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 __author__ = 'Destiny Peterson & Matthew Walker'
-__date__ = '2023-11-11'
-__version__ = '0.46.3'
+__date__ = '2023-11-12'
+__version__ = '0.47.0'
 
 __all__ = [
     # Modules: